[ADD] adding the table app to the application layer; adding the base table that automatically generates the context table dictionary to assemble a bootstrap table automatically; adding the organisms and organism corresponding tables
