+++ src/application/tables/__init__.py
@@ -1,0 +1,5 @@
+from .organisms import (OrganismsTable,
+                        OrganismRegulatorsTable,
+                        OrganismGenesTable,
+                        OrganismBindingsTable,
+                        OrganismInteractionsTable)
